requirements.txt

beautifulsoup4==4.11.1
certifi==2020.6.20
charset-normalizer==2.1.1
click==7.1.2
docopt==0.6.2
Flask==1.1.2
gunicorn==20.1.0
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.2
pipreqs==0.4.11
requests==2.28.1
scikit-learn==0.23.2
scipy==1.5.2
sklearn==0.0
soupsieve==2.3.2.post1
threadpoolctl==2.1.0
urllib3==1.26.13
Werkzeug==1.0.1
wincertstore==0.2
yarg==0.1.9
mysql-connector-python==8.0.31